Phase 3.4-3.5: Create CLI package and move command functions
Phase 3.4: CLI Package Structure
- Create doc_search/cli/ package directory
- Create cli/__init__.py with main() entry point
- Create cli/parsers.py with argument parser definitions

Phase 3.5: Move Command Functions
- Create cli/commands.py with all cmd_* implementations:
  - cmd_crawl: Crawl documentation sites
  - cmd_index: Build search index
  - cmd_search: Search the index
  - cmd_autocomplete: Get autocomplete suggestions
  - cmd_interactive: Interactive search mode
  - cmd_stats: Show site statistics
  - cmd_list: List crawled sites
  - cmd_serve: Start web UI server

- Simplify __main__.py to just import and call cli.main()
- Helper functions (get_site_dir, get_auth) moved to commands.py
- DEFAULT_DATA_DIR constant moved to commands.py

All 427 tests pass. CLI works correctly.
